Flytt søkeboks til venstre margin
