+++ src/App/Hovedside/Sak/Saksfilter/Saksfilter.tsx
@@ -13,15 +13,20 @@ import {
     Virksomhetsmeny
 } from "./Virksomhetsmeny/Virksomhetsmeny";
 import {byggOrganisasjonstre} from "./ByggOrganisasjonstre";
+import { Søkeboks } from './Søkeboks';
+import { Filter } from '../Saksoversikt/useOversiktStateTransitions';
 
 
 type SaksfilterProps = {
+    filter: Filter;
+    setFilter: (filter: Filter) => void;
     valgteVirksomheter: Organisasjon[] | "ALLEBEDRIFTER";
     setValgteVirksomheter: (valgteVirksomheter: Organisasjon[] | "ALLEBEDRIFTER") => void;
     organisasjoner: Organisasjon[];
 }
 
-export const Saksfilter = ({valgteVirksomheter, setValgteVirksomheter, organisasjoner}: SaksfilterProps) => {
+
+export const Saksfilter = ({valgteVirksomheter, setValgteVirksomheter, organisasjoner, filter, setFilter}: SaksfilterProps) => {
     const [organisasjonstre, setOrganisasjonstre] = useState<OrganisasjonEnhet[]>()
 
     useEffect(() => {
@@ -39,33 +44,7 @@ export const Saksfilter = ({valgteVirksomheter, setValgteVirksomheter, organisas
                          valgteEnheter={valgteVirksomheter}
                          settValgteEnheter={setValgteVirksomheter}/>
 
-        <div className="saksfilter_søk-sak">
-            <BodyShort className="saksfilter_headers">Søk blant saker</BodyShort>
-            <Search label="Søk alle NAV sine sider" variant="secondary"/>
-        </div>
-
-        <div className="saksfilter_type-sak">
-            <CheckboxGroup
-                legend={<BodyShort className="saksfilter_headers">Type saker</BodyShort>}
-                onChange={(val: any[]) => {
-                    null
-                }}
-            >
-                <div className={'saksfiler_type-sak_valg_'}>
-                    <Checkbox value="Fritak">Fritak arbeidsgiverperiode (4)</Checkbox>
-                    <Checkbox value="Korona">Korona sykdom (10)</Checkbox>
-                    <Checkbox value="Lønnskomp">Lønnskompensasjon (23)</Checkbox>
-                    <Checkbox value="Refusjon">Refusjon arbeidgsiverperioden (26)</Checkbox>
-                    <Checkbox value="Utestengt">Utestengt EØS borger (3)</Checkbox>
-                    <Checkbox value="Permittering">Permittering og nedbemmaning (14)</Checkbox>
-                </div>
-            </CheckboxGroup>
-        </div>
-
-        <Select label="Periode" hideLabel
-                defaultValue={12}>
-            <option value={12}> Siste 12 måneder</option>
-        </Select>
+        <Søkeboks filter={filter} byttFilter={setFilter}></Søkeboks>
     </div>;
 }
