+++ src/App/Hovedside/Sak/Saksfilter/Søkeboks.tsx
@@ -1,0 +1,37 @@
+import { BodyShort, Search } from '@navikt/ds-react';
+import React, { useEffect, useRef } from 'react';
+import { Filter } from '../Saksoversikt/useOversiktStateTransitions';
+import './Saksfilter.css'
+
+export type SøkeboksProps = {
+    filter: Filter,
+    byttFilter: (filter: Filter) => void;
+}
+
+export const Søkeboks = ({filter, byttFilter}: SøkeboksProps) => {
+    const searchElem = useRef<HTMLInputElement>(null)
+
+    useEffect(() => {
+        if (searchElem.current) {
+            searchElem.current.value = filter.tekstsoek
+        }
+    }, [searchElem.current, filter.tekstsoek])
+
+    return <form
+        className="saksfilter_søk-sak"
+        onSubmit={(e) => {
+            e.preventDefault()
+            const tekstsoek = searchElem?.current?.value ?? ''
+            byttFilter({...filter, tekstsoek})
+        }}
+    >
+        <BodyShort className="saksfilter_headers">Søk blant saker</BodyShort>
+        <Search
+            label="Søk blandt saker"
+            size="medium"
+            variant="secondary"
+            ref={searchElem}
+        />
+    </form>
+}
+
